Modularização do código e implementação do sistema de automação
- Alterado  para centralizar a execução do bot e o menu interativo
- Criado  para gerenciar a automação e agendamentos
- Implementado sistema de automação que roda toda segunda-feira
- Adicionado o sistema de threads para rodar a automação em segundo plano sem bloquear o menu
- Agora é possível ativar, desativar e configurar a automação diretamente pelo terminal
- Menu atualizado para permitir alteração dos parâmetros da automação (termo de busca, localização, mensagem e número de páginas)
