+++ config.json
@@ -1,0 +1,9 @@
+{
+    "automation_enabled": false,
+    "last_execution": null,
+    "schedules": [],
+    "termo_busca": "Desenvolvedor de Software",
+    "localizacao": "",
+    "mensagem_personalizada": "",
+    "max_pages": 10
+}
